refactor: add core package
