+++ bomplatform-core/src/main/java/com/javabom/bomplatform/core/mission/model/MissionState.java
@@ -1,0 +1,5 @@
+package com.javabom.bomplatform.core.mission.model;
+
+public enum  MissionState {
+    OPEN, CLOSE
+}
