+++ bomplatform-core/src/main/java/com/javabom/bomplatform/core/mission/model/MissionStep.java
@@ -1,4 +1,4 @@
-package com.javabom.bomplatform.mission.model;
+package com.javabom.bomplatform.core.mission.model;
 
 public enum MissionStep {
     FIRST, SECOND, THIRD
