+++ bomplatform-core/src/main/java/com/javabom/bomplatform/core/mission/repository/MissionRepository.java
@@ -1,0 +1,9 @@
+package com.javabom.bomplatform.core.mission.repository;
+
+import com.javabom.bomplatform.core.mission.model.Mission;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MissionRepository extends JpaRepository<Mission, Long> {
+}
