+++ bomplatform-core/src/main/java/com/javabom/bomplatform/core/progressmission/model/ProgressMissionState.java
@@ -1,4 +1,4 @@
-package com.javabom.bomplatform.progressmission.model;
+package com.javabom.bomplatform.core.progressmission.model;
 
 public enum ProgressMissionState {
     OPEN, REVIEW, CLOSE
