+++ bomplatform-core/src/main/java/com/javabom/bomplatform/core/review/repository/ReviewRepository.java
@@ -1,0 +1,9 @@
+package com.javabom.bomplatform.core.review.repository;
+
+import com.javabom.bomplatform.core.review.model.Review;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ReviewRepository extends JpaRepository<Review, Long> {
+}
