+++ bomplatform-core/src/main/java/com/javabom/bomplatform/core/user/model/UserRole.java
@@ -1,4 +1,4 @@
-package com.javabom.bomplatform.user.model;
+package com.javabom.bomplatform.core.user.model;
 
 public enum UserRole {
     CHALLENGER, REVIEWER
