+++ bomplatform-core/src/main/java/com/javabom/bomplatform/core/user/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package com.javabom.bomplatform.core.user.repository;
+
+import com.javabom.bomplatform.core.user.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+}
