--- bomplatform-core/src/main/java/com/javabom/bomplatform/mission/model/MissionState.java
@@ -1,5 +1,0 @@
-package com.javabom.bomplatform.mission.model;
-
-public enum  MissionState {
-    OPEN, CLOSE
-}
