--- bomplatform-core/src/main/java/com/javabom/bomplatform/mission/repository/MissionRepository.java
@@ -1,7 +1,0 @@
-package com.javabom.bomplatform.mission.repository;
-
-import com.javabom.bomplatform.mission.model.Mission;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MissionRepository extends JpaRepository<Mission, Long> {
-}
