--- bomplatform-core/src/main/java/com/javabom/bomplatform/progressmission/repository/ProgressMissionRepository.java
@@ -1,7 +1,0 @@
-package com.javabom.bomplatform.progressmission.repository;
-
-import com.javabom.bomplatform.progressmission.model.ProgressMission;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProgressMissionRepository extends JpaRepository<ProgressMission, Long> {
-}
