--- bomplatform-core/src/main/java/com/javabom/bomplatform/review/repository/ReviewRepository.java
@@ -1,7 +1,0 @@
-package com.javabom.bomplatform.review.repository;
-
-import com.javabom.bomplatform.review.model.Review;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ReviewRepository extends JpaRepository<Review, Long> {
-}
