--- bomplatform-core/src/main/java/com/javabom/bomplatform/user/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package com.javabom.bomplatform.user.repository;
-
-import com.javabom.bomplatform.user.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-}
